requirements.txt
opencv-python
mediapipe
numpy
pytesseract
Pillow